feat: Added authorization checks

+++ controllers/feed.js
@@ -1,8 +1,9 @@
 const fs = require("fs");
 const path = require("path");
-
+const User = require("../models/user");
 const { validationResult } = require("express-validator");
 const Post = require("../models/post");
+
 exports.getPosts = (req, res, next) => {
   const currentPage = req.query.page || 1;
   const perPage = 2;
@@ -44,18 +45,28 @@ exports.postCreatePost = (req, res, next) => {
   }
   const imageUrl = req.file.path;
   const { title, content } = req.body;
+  let creator;
   const post = new Post({
     title,
     content,
-    creator: { name: "rDev" },
+    creator: req.userId,
     imageUrl,
   });
   post
     .save()
     .then((post) => {
+      return User.findById(req.userId);
+    })
+    .then((user) => {
+      creator = user;
+      user.posts.push(post);
+      return user.save();
+    })
+    .then((result) => {
       res.status(201).json({
         message: "Post created successfully!",
         post,
+        creator,
       });
     })
     .catch((error) => {
@@ -110,6 +121,11 @@ exports.putEditPost = (req, res, next) => {
         error.statusCode = 404;
         throw error;
       }
+      if (post.creator.toString !== req.userId) {
+        const error = new Error("Not authorized!!!");
+        error.statusCode = 403;
+        throw error;
+      }
       if (imageUrl !== post.imageUrl) {
         clearImage(post.imageUrl);
       }
@@ -138,12 +154,23 @@ exports.deletePost = (req, res, next) => {
         error.statusCode = 404;
         throw error;
       }
+      if (post.creator.toString() !== req.userId) {
+        const error = new Error("Not authorized!!!");
+        error.statusCode = 403;
+        throw error;
+      }
       // Check logged in user
       clearImage(post.imageUrl);
       return Post.findByIdAndRemove(postId);
     })
     .then((result) => {
-      console.log("delete result :", result);
+      return User.findById(req.userId);
+    })
+    .then((user) => {
+      user.posts.pull(postId);
+      return user.save();
+    })
+    .then((user) => {
       res.status(200).json({ message: "Post deleted successfully !!" });
     })
     .catch((error) => {
